Added backwards compatible parsing
